feat: S3 timeout and resilience support (P0 + P1) poll()-based I/O with socket timeouts and cancel support: - Replace blocking read()/write() with poll()-based wrappers checking InterruptPending; - Add GUC yezzey.yproxy_socket_timeout (default 120s); - Add ereport(ERROR) in all I/O error paths instead of silent failures; - Fix poll() POLLIN/POLLHUP ordering to avoid data loss on simultaneous events; Error propagation (P1): - Propagate network errors in yproxy_lister instead of returning empty vector; - ereport(ERROR) on truncated metadata in readObjectMetaBody; - Write path error propagation with ereport(ERROR) in storage.cpp; - Reorder io_close() before metadata update in offload path;

+++ include/storage.h
@@ -53,7 +53,7 @@ EXTERNC int statRelationSpaceUsage(Relation aorel, int segno, int64 modcount,
                                    size_t *local_commited_bytes,
                                    size_t *external_bytes);
 
-EXTERNC int statExternalTotal(Relation aorel, int segindx);
+EXTERNC int64_t statExternalTotal(Relation aorel, int segindx);
 
 EXTERNC int statRelationChunksSpaceUsage(Relation aorel, size_t *local_bytes,
                                          size_t *local_commited_bytes,
